requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dimod==0.11.1
diskcache==5.4.0
dwave-cloud-client==0.10.1
dwave-greedy==0.2.5
dwave-hybrid==0.6.7
dwave-inspector==0.2.11
dwave-inspectorapp==0.2.4
dwave-neal==0.5.9
dwave-networkx==0.8.12
dwave-ocean-sdk==5.1.0
dwave-preprocessing==0.4.0
dwave-system==1.15.0
dwave-tabu==0.4.5
dwavebinarycsp==0.2.0
entrypoints==0.4
executing==0.8.3
fasteners==0.17.3
Flask==2.1.2
fonttools==4.34.4
graphviz==0.20.1
homebase==1.0.1
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.13.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.1
jupyter-core==4.10.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
minorminer==0.2.8.post1
nest-asyncio==1.5.5
networkx==2.8.8
numpy==1.22.4
packaging==21.3
parso==0.8.3
penaltymodel==1.0.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plucky==0.4.3
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.1
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pyzmq==23.1.0
rectangle-packer==2.0.1
requests==2.27.1
scipy==1.8.1
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.2.2.post1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
